split repeated Pose and Armature classes to new file
